Add state changes and music
Added the logic for a worker thread that gets spawned when the robot is instantiated. This second thread will play a background music for the game. A channel is also established for the robot to tell the thread when to play state-change sounds. Code for the robot to be in different states is also added.